Fix midnight error
Because of summer and winter saving time, right now when it is between 0:00am and 1:00am the behaviour is buggy. This is because technically a full day will have passed only after 1am (because of saving time). This meant that the word to be guessed was still for the previous day, but the game was loaded for the new day